Move higher level parsing functions to datafile.c
