+++ src/init.c
@@ -28,6 +28,7 @@
 #include "cave.h"
 #include "cmds.h"
 #include "cmd-core.h"
+#include "datafile.h"
 #include "effects.h"
 #include "game-event.h"
 #include "generate.h"
@@ -51,7 +52,6 @@
 #include "obj-util.h"
 #include "object.h"
 #include "option.h"
-#include "parser.h"
 #include "player.h"
 #include "player-history.h"
 #include "player-quest.h"
