+++ src/mon-init.h
@@ -31,7 +31,5 @@ extern struct file_parser mon_base_parser;
 extern struct file_parser pit_parser;
 extern struct file_parser pain_parser;
 
-void write_flags(ang_file *fff, const char *intro_text, bitflag *flags,
-				 int flag_size, const char *names[]);
 
 #endif /* MONSTER_INIT_H_ */
