+++ src/parser.h
@@ -47,16 +47,6 @@ struct parser_state {
 	char *msg;
 };
 
-struct file_parser {
-	const char *name;
-	struct parser *(*init)(void);
-	errr (*run)(struct parser *p);
-	errr (*finish)(struct parser *p);
-	void (*cleanup)(void);
-};
-
-extern const char *parser_error_str[PARSE_ERROR_MAX];
-
 extern struct parser *parser_new(void);
 extern enum parser_error parser_parse(struct parser *p, const char *line);
 extern void parser_destroy(struct parser *p);
@@ -74,24 +64,5 @@ extern struct random parser_getrand(struct parser *p, const char *name);
 extern wchar_t parser_getchar(struct parser *p, const char *name);
 extern int parser_getstate(struct parser *p, struct parser_state *s);
 extern void parser_setstate(struct parser *p, unsigned int col, const char *msg);
-
-errr run_parser(struct file_parser *fp);
-errr parse_file_quit_not_found(struct parser *p, const char *filename);
-errr parse_file(struct parser *p, const char *filename);
-void cleanup_parser(struct file_parser *fp);
-int lookup_flag(const char **flag_table, const char *flag_name);
-errr grab_rand_value(random_value *value, const char **value_type,
-					 const char *name_and_value);
-errr grab_int_value(int *value, const char **value_type,
-					const char *name_and_value);
-errr grab_index_and_int(int *value, int *index, const char **value_type,
-						const char *prefix, const char *name_and_value);
-errr grab_base_and_int(int *value, char **base, const char *name_and_value);
-errr grab_name(const char *from, const char *what, const char *list[], int max,
-			   int *num);
-errr grab_flag(bitflag *flags, const size_t size, const char **flag_table,
-			   const char *flag_name);
-errr remove_flag(bitflag *flags, const size_t size, const char **flag_table,
-				 const char *flag_name);
 
 #endif /* !PARSER_H */
